requirements.txt
openai>=1.30.0
PyYAML>=6.0
pyarrow>=14.0.0
pytest>=8.0.0
